+++ torch_geometric/nn/dense/linear.py
@@ -198,12 +198,13 @@ class HeteroLinear(torch.nn.Module):
             :obj:`type_vec` is sorted. This avoids internal re-sorting of the
             data and can improve runtime and memory efficiency.
             (default: :obj:`False`)
-        use_segmm (bool, optional): If set to :obj:`True` and pyg-lib is
+        use_segmm (bool, optional): If set to :obj:`True` and :obj:`pyg-lib` is
             installed, this module will use the fused :obj:`segment_matmul`
             kernel to parallelize the linear transformation across types. If
             set to :obj:`False`, :obj:`segment_matmul` will not be used. If
-            left as :obj:`None` and pyg-lib is installed, the module will
-            determine heuristically whether to use :obj:`segment_matmul`.
+            left as :obj:`None` and :obj:`pyg-lib` is installed, the module
+            will determine heuristically whether to use :obj:`segment_matmul`.
+            (default: :obj:`None`)
         **kwargs (optional): Additional arguments of
             :class:`torch_geometric.nn.Linear`.
 
@@ -213,18 +214,23 @@ class HeteroLinear(torch.nn.Module):
           type vector :math:`(*)`
         - **output:** features :math:`(*, F_{out})`
     """
-    def __init__(self, in_channels: int, out_channels: int, num_types: int,
-                 is_sorted: bool = False, use_segmm: Optional[bool] = None,
-                 **kwargs):
+    def __init__(
+        self,
+        in_channels: int,
+        out_channels: int,
+        num_types: int,
+        is_sorted: bool = False,
+        use_segmm: Optional[bool] = None,
+        **kwargs,
+    ):
         super().__init__()
 
         self.in_channels = in_channels
         self.out_channels = out_channels
         self.num_types = num_types
         self.is_sorted = is_sorted
+        self.use_segmm = use_segmm
         self.kwargs = kwargs
-
-        self.use_segmm: int = int(use_segmm) if use_segmm is not None else -1
 
         if self.in_channels == -1:
             self.weight = nn.parameter.UninitializedParameter()
@@ -253,7 +259,7 @@ class HeteroLinear(torch.nn.Module):
             type_vec (torch.Tensor): A vector that maps each entry to a type.
         """
         if (torch_geometric.typing.WITH_PYG_LIB
-                and (self.use_segmm == -1 or bool(self.use_segmm))):
+                and (self.use_segmm is None or self.use_segmm)):
             assert self.weight is not None
 
             perm: Optional[Tensor] = None
@@ -263,7 +269,7 @@ class HeteroLinear(torch.nn.Module):
                     x = x[perm]
 
             type_vec_ptr = index2ptr(type_vec, self.num_types)
-            if self.use_segmm == -1:
+            if self.use_segmm is None:
                 self.use_segmm = segmatmul_heuristic(x, type_vec_ptr,
                                                      self.weight)
             out = pyg_lib.ops.segment_matmul(x, type_vec_ptr, self.weight)
@@ -322,6 +328,13 @@ class HeteroDictLinear(torch.nn.Module):
         out_channels (int): Size of each output sample.
         types (List[Any], optional): The keys of the input dictionary.
             (default: :obj:`None`)
+        use_segmm (bool, optional): If set to :obj:`True` and :obj:`pyg-lib` is
+            installed, this module will use the fused :obj:`segment_matmul`
+            kernel to parallelize the linear transformation across types. If
+            set to :obj:`False`, :obj:`segment_matmul` will not be used. If
+            left as :obj:`None` and :obj:`pyg-lib` is installed, the module
+            will determine heuristically whether to use :obj:`segment_matmul`.
+            (default: :obj:`None`)
         **kwargs (optional): Additional arguments of
             :class:`torch_geometric.nn.Linear`.
     """
@@ -330,6 +343,7 @@ class HeteroDictLinear(torch.nn.Module):
         in_channels: Union[int, Dict[Any, int]],
         out_channels: int,
         types: Optional[Any] = None,
+        use_segmm: Optional[bool] = None,
         **kwargs,
     ):
         super().__init__()
@@ -359,6 +373,7 @@ class HeteroDictLinear(torch.nn.Module):
 
         self.in_channels = in_channels
         self.out_channels = out_channels
+        self.use_segmm = use_segmm
         self.kwargs = kwargs
 
         self.lins = torch.nn.ModuleDict({
@@ -386,8 +401,13 @@ class HeteroDictLinear(torch.nn.Module):
 
         # Only apply fused kernel for more than 10 types, otherwise default
         # back to sequential computation (which is faster for these cases).
+        if self.use_segmm is None:
+            use_segmm = len(x_dict) >= 10
+        else:
+            self.use_segmm = use_segmm
+
         if (torch_geometric.typing.WITH_GMM and not torch.jit.is_scripting()
-                and len(x_dict) >= 10):
+                and self.use_segmm):
             xs, weights, biases = [], [], []
             for key, lin in self.lins.items():
                 if key in x_dict: